+++ spec/paper_trail_spec.rb
@@ -94,36 +94,54 @@ RSpec.describe PaperTrail do
   end
 
   describe ".version" do
-    it { expect(described_class).to respond_to(:version) }
-    it { expect(described_class.version).to eq(described_class::VERSION::STRING) }
+    it "returns the expected String" do
+      expect(described_class.version).to eq(described_class::VERSION::STRING)
+    end
   end
 
-  describe ".whodunnit" do
-    context "with block passed" do
-      it "sets whodunnit only for the block passed" do
-        described_class.whodunnit("foo") do
-          expect(described_class.whodunnit).to eq("foo")
-        end
+  describe "deprecated methods" do
+    before do
+      allow(ActiveSupport::Deprecation).to receive(:warn)
+    end
 
-        expect(described_class.whodunnit).to be_nil
-      end
-
-      it "sets whodunnit only for the current thread" do
-        described_class.whodunnit("foo") do
-          expect(described_class.whodunnit).to eq("foo")
-          Thread.new { expect(described_class.whodunnit).to be_nil }.join
-        end
-
-        expect(described_class.whodunnit).to be_nil
+    shared_examples "it delegates to request" do |method, args|
+      it do
+        arguments = args || [no_args]
+        allow(described_class.request).to receive(method)
+        described_class.public_send(method, *args)
+        expect(described_class.request).to have_received(method).with(*arguments)
+        expect(ActiveSupport::Deprecation).to have_received(:warn)
       end
     end
 
-    context "when set to a proc" do
-      it "evaluates the proc each time a version is made" do
-        call_count = 0
-        described_class.whodunnit = proc { call_count += 1 }
-        expect(described_class.whodunnit).to eq(1)
-        expect(described_class.whodunnit).to eq(2)
+    it_behaves_like "it delegates to request", :clear_transaction_id, nil
+    it_behaves_like "it delegates to request", :enabled_for_controller=, [true]
+    it_behaves_like "it delegates to request", :enabled_for_model, [Widget, true]
+    it_behaves_like "it delegates to request", :enabled_for_model?, [Widget]
+    it_behaves_like "it delegates to request", :whodunnit=, [:some_whodunnit]
+    it_behaves_like "it delegates to request", :whodunnit, nil
+    it_behaves_like "it delegates to request", :controller_info=, [:some_whodunnit]
+    it_behaves_like "it delegates to request", :controller_info, nil
+    it_behaves_like "it delegates to request", :transaction_id=, 123
+    it_behaves_like "it delegates to request", :transaction_id, nil
+
+    describe "whodunnit with block" do
+      it "delegates to request" do
+        allow(described_class.request).to receive(:with)
+        described_class.whodunnit(:some_whodunnit) { :some_block }
+        expect(ActiveSupport::Deprecation).to have_received(:warn)
+        expect(described_class.request).to have_received(:with) do |*args, &block|
+          expect(args).to eq([{ whodunnit: :some_whodunnit }])
+          expect(block.call).to eq :some_block
+        end
+      end
+    end
+
+    describe "whodunnit with invalid arguments" do
+      it "raises an error" do
+        expect { described_class.whodunnit(:some_whodunnit) }.to raise_error(ArgumentError) do |e|
+          expect(e.message).to eq "Invalid arguments"
+        end
       end
     end
   end